handled error

+++ API/delete.js
@@ -5,6 +5,10 @@ const db = require('../db');
 Router.delete('/', (req, res) => {
     const { rno } = req.body;
 
+    if (!rno) {
+        return res.status(400).json({ message: ' rno not found' });
+    }
+
     const sql = `DELETE FROM students WHERE rno=?`;
     db.run(sql, [rno], (err) => {
         if (err) {
